+++ src/app/[jobBoard]/(auth)/sign-in/page.tsx
@@ -1,0 +1,5 @@
+import SignIn from "./sign-in";
+
+export default function SignInPage() {
+  return <SignIn />;
+}
