+++ src/app/[jobBoard]/(auth)/sign-in/sign-in.tsx
@@ -107,6 +107,7 @@ export default function SignIn() {
                 },
                 onSuccess: async () => {
                   router.replace("/dashboard");
+                  router.refresh();
                 },
               }
             );
