+++ src/app/[jobBoard]/(marketing)/about/about.mdx
