+++ src/app/[jobBoard]/(marketing)/pricing/pricing-inputs.tsx
