+++ src/app/[jobBoard]/dashboard/account/page.tsx
