+++ src/app/[jobBoard]/dashboard/page.tsx
