+++ src/app/[jobBoard]/dashboard/posts/[id]/components/hide-post.tsx
