+++ src/app/[jobBoard]/dashboard/posts/[id]/edit/edit-post-form.tsx
@@ -30,7 +30,7 @@ import {
 } from "@/app/lib/constants";
 import { Textarea } from "@/components/ui/textarea";
 import Link from "next/link";
-import { updateJobPost } from "@/app/admin/dashboard/posts/create/actions";
+import { updateJobPost } from "@/app/[jobBoard]/dashboard/posts/create/actions";
 import FormSubmitButton from "@/components/form-submit-button";
 import { useParams, useRouter } from "next/navigation";
 import type { EditJobPosting } from "@/app/lib/job-postings/schema";
