+++ src/app/[jobBoard]/dashboard/posts/[id]/edit/page.tsx
