+++ src/app/[jobBoard]/dashboard/posts/[id]/page.tsx
