+++ src/app/[jobBoard]/dashboard/posts/create/page.tsx
