+++ src/app/[jobBoard]/page.tsx
@@ -33,7 +33,7 @@ export default async function Page({
   });
 
   return (
-    <main>
+    <>
       <div className="container mx-auto space-y-4 pt-4 text-primary">
         <Form action={`/`} className="flex gap-2">
           <Input name="jobTitle" placeholder="Search Jobs..." />
@@ -74,6 +74,6 @@ export default async function Page({
           <P className="text-center">No jobs matched the filter.</P>
         )}
       </section>
-    </main>
+    </>
   );
 }
