--- src/app/admin/(marketing)/privacy-policy/page.tsx
@@ -1,10 +1,0 @@
-import PrivacyPolicy from "@/app/admin/(marketing)/privacy-policy/privacy-policy.mdx";
-import { mdxComponents } from "@/app/admin/(marketing)/components/mdx";
-
-export default async function Page() {
-  return (
-    <main className="mx-auto max-w-4xl py-20">
-      <PrivacyPolicy components={mdxComponents} />
-    </main>
-  );
-}
