--- src/app/admin/(marketing)/terms-of-service/page.tsx
@@ -1,10 +1,0 @@
-import TermsOfService from "@/app/admin/(marketing)/terms-of-service/terms-of-service.mdx";
-import { mdxComponents } from "@/app/admin/(marketing)/components/mdx";
-
-export default async function Page() {
-  return (
-    <main className="mx-auto max-w-4xl py-20">
-      <TermsOfService components={mdxComponents} />
-    </main>
-  );
-}
