--- src/app/admin/dashboard/layout.tsx
@@ -1,20 +1,0 @@
-import Navbar from "@/components/navbar";
-import type { Metadata } from "next";
-
-export const metadata: Metadata = {
-  title: "Manage Opportunities",
-  description: "Streamlining the LMIA process for Canadian employers",
-};
-
-export default async function AdminLayout({
-  children,
-}: Readonly<{
-  children: React.ReactNode;
-}>) {
-  return (
-    <>
-      <Navbar />
-      {children}
-    </>
-  );
-}
