--- src/app/admin/sign-in/page.tsx
@@ -1,9 +1,0 @@
-import SignIn from "./sign-in";
-
-export default function SignInPage() {
-  return (
-    <main className="h-dvh content-center px-8">
-      <SignIn />
-    </main>
-  );
-}
