+++ src/app/lib/constants.ts
@@ -13,7 +13,6 @@ export const JOB_BOARD_TITLES: Record<JobBoard, string> = {
 };
 
 export const JOB_BOARD_DOMAINS: Record<string, JobBoard | "admin"> = {
-  "manageopportunities.ca": "admin",
   "allopportunities.ca": "all",
   "accessibleopportunities.ca": "accessible",
   "asylumopportunities.ca": "asylum",
