+++ src/components/navbar.tsx
@@ -5,35 +5,24 @@ import { headers } from "next/headers";
 import { Button } from "./ui/button";
 import { SignOut } from "./sign-out";
 
-export default async function Navbar({ title }: { title?: string | null }) {
+const links = [
+  { text: "Home", url: "/" },
+  { text: "Pricing", url: "/pricing" },
+  { text: "Contact Us", url: "/contact-us" },
+];
+
+export default async function Navbar({ title }: { title: string }) {
   const data = await auth.api.getSession({
     headers: await headers(),
   });
 
-  const links = [
-    { text: "Home", url: data ? "/dashboard" : "/" },
-    { text: "Pricing", url: "/pricing" },
-    { text: "Contact Us", url: "/contact-us" },
-  ];
-
   return (
     <header className="flex h-16 flex-col items-center bg-white px-4 sm:h-14 sm:flex-row lg:px-6">
-      {!title ? (
-        <Link
-          className="flex items-center justify-center"
-          href={data ? "/dashboard" : "/"}>
-          <span className="text-xl font-bold tracking-tighter">
-            Manage Opportunities
-          </span>
-        </Link>
-      ) : (
-        <Link href={`/`}>
-          <span className="text-xl font-bold tracking-tighter text-primary">
-            {title}
-          </span>
-        </Link>
-      )}
-
+      <Link className="flex items-center justify-center" href={"/"}>
+        <span className="text-xl font-bold tracking-tighter text-primary">
+          {title}
+        </span>
+      </Link>
       <nav className="flex items-center justify-center gap-4 py-2 sm:ml-auto sm:mt-0 sm:gap-6">
         {links.map((link) => (
           <Button key={link.text} asChild variant="link">
@@ -44,8 +33,12 @@ export default async function Navbar({ title }: { title?: string | null }) {
         {data ? (
           <>
             <Button asChild variant="link">
+              <Link href="/dashboard">Dashboard</Link>
+            </Button>
+            <Button asChild variant="link">
               <Link href="/dashboard/account">Account</Link>
             </Button>
+
             <SignOut />
           </>
         ) : (
