+++ src/components/sign-out.tsx
@@ -13,6 +13,7 @@ export function SignOut() {
           fetchOptions: {
             onSuccess: () => {
               router.push("/");
+              router.refresh();
             },
           },
         })
